Adjust reference point for pkg. env. savings calcs
When envelope savings are calculated for an HVAC/envelope package, it’s relative to all possible heating/cooling markets that the envelope measures in the analysis could affect and which overlap with the HVAC measure markets. Rather than tying the denominator of that calculation to the envelope measures in the set, which may not always cover all possible heating/cooling markets, it should be relative to a predetermined total for heating/cooling across all the markets that overlap with the HVAC measures. This commit makes that change.

This commit also tweaks the end use matching between envelope and HVAC heating markets, which had previously been including secondary heating markets for envelope in the match with heating markets for HVAC.

Finally, initialize overlapping envelope/HVAC savings fractions as 0 and relative performance as 1, rather than initializing both as 1.