Update start year, fuel switching to resistance, reporting
The start year was hardcoded in the preparation module and supporting heating/cooling totals data to 2024. For fuel switching measures that convert to electric resistance, keys for the switched to technology remain identical to the baseline for the cooling end use (e.g., if it's a NG + central AC baseline, switched to mseg information would indicate electric resistance heater + central AC). The YML files were updated such that the run module always reports out stock information across scenarios in the BSS batch run.