Add check on secondary heating in measure def
To avoid accounting issues, use convention that a measure that pairs primary and secondary heating must match the fuel of the latter with the former. Without this, measures that pair fossil-based heating FS with electric cooling inadvertently pull in secondary electric heating along with secondary gas heating (the two have identical tech names). That secondary electric heating is tagged as occuring in a fuel switching context and is not correctly competed with secondary electric heating that is not in a fuel switching context.