Support modifications of AEO incentive levels
Create and process new data (./scout/supporting_data/sub_fed/incentives.csv) to allow users to remove, extend, or replace AEO incentive levels for specific msegs in the Scout forecast.

A new function, import_state_data, was written in ecm_prep to read in and finalize the inputs for incentives modifications and (in the future) modifications to rates in one place.

Additionally, “federal” and “non-federal” branches on the existing residential incentives levels were added that split out the previously summed incentives across these dimensions, to allow more flexibility in making modifications to federal vs. sub-federal residential incentive levels for the residential sector (commercial incentives are not split out in this way in the AEO data.)

Users may modify incentive levels for a given applicable baseline segment or segment that a measure is switching to in three ways:

* Remove: exclude the AEO incentive from consideration beginning in a given start year (or for all years).
* Replace: if there is no removal (first bullet), specify a custom federal or state-level incentive that replaces any AEO incentive level previously specified for the given segment.
* Extend: if there is no removal or replacement (first and second bullets), find the largest level of the AEO incentive in years prior to a given start year and extend it through a given end year, possibly with a multiplier added on top of the existing AEO incentive.

For replacements of existing incentives for fuel switching, a user input determines whether the incentive allows the original fuel equipment to be retained as backup. For replacements, users must also specify information about how to qualify for the incentive – a performance threshold and units and the incentive level itself, which is either a % credit on installed cost or rebate $ decremented from installed cost.

Users may specify modifications to federal and sub-federal incentives separately, or can specify a modification that applies to both.

For sub-state-level incentives modifications, users may specify an applicable fraction that downscales the modification’s impact to only the portion of a state’s energy use that is affected by a sub-state policy.

An additional reference case measure flag, “ref_case_flag” was needed to ensure that incentives process correctly for add-on reference case measures (all add-on measures were previously excluded to prevent attempts to process incentives for, e.g., controls measures with relative % savings units, however ref. case measures are often defined as add-ons with zero performance or cost additions to the baseline and incentives should be assigned to these measures). The measure set was regenerated with this flag reflected.

Addresses issue #410.