Fixes to incentives calculations
Ensure that when users impose heating incentives for HPs, any existing incentives for cooling msegs are zeroed out. Ensure that suppression of incentives (e.g., federal or state only, etc.) is consistent in use of strings to identify incentives that should be filtered out. Reflect that incentives use warm climate performance settings for dual fuel HPs, even when the HP is installed in a cold climate.