Correct issues with tech-level processing
The tech-level factors were not actually being applied. Pull in the correct tech flag for keying in the tech-resolved factors and apply it to the rows to get the right factors. Use an explicit mapping between Scout technology names and the names in the disaggregation files to pull in the data.