Add capability to read/write endogenous conversion rates
Users may specify the 'write_hp_conv_fracs' option in run to calculate rates of conversion from fossil to electric equipment and/or to new equipment classes (across fuels or within the same fuel, such as resistance to HPs). These rates will be written as new scenario data to the existing 'hp_convert_rates.json' file, alongside the legacy Guidehouse rates. The rates will not be written out in cases where measure data were prepared using exogenous rate scenarios; only endenously-determined results will be processed in this way.

The ecm_prep module is modified to be able to distinguish between the legacy rates (prepended 'gh') and the endogenously-determined rate scenarios, which is necessary because the latter has a different format, where the data are broken out by whatever Scout regions were used in the endogenous runs (rather than by Census Region as in the Guidehouse data) and go down to the end-use-level (rather than the technology-level as in the Guidehouse data).