Reconfigure reporting of linked stock/energy costs
Users may now allow linked stock costs to be ignored in the calculations that determine technology adoption, but reported nonetheless under total capital costs for a measure/scenario. Linked energy costs are now always reported, even if they are ignored in the calculations that determine technology adoption, to ensure a full accounting of the costs of the energy consumed in each simulation, which ensures comparability across scenario costs.

Additionally, a bug in the variable used to record and report remaining baseline fuel use for dual fuel measures was fixed.