Add unit tests for dual fuel and added cooling
Add two tests,  and  to  class in . These tests engage the dual fuel measure and no baseline cooling -> HP measure capabilities that were added via PRs #530 and #531, respectively. Both features are tested for a furnace to HP conversion case for example CA residential microsegments.

In addition, a small change to the reporting of efficient-captured energy use from added cooling was implemented in ecm_prep to ensure that efficient-captured energy is being reported before trying to calculate it for the added cooling case.